Add HelloWorld_NKK to module
The NKK is slightly different from the normal HelloWorld. Perhaps it would be
better to combine the two to a single file and have a small switch to select
which one the player wants to use.

The diffence in keyboard mapping is handled by the CodeKlavier/Mapping package.
This allows easy reuse or changing of the mapping.